+++ src/js/addons/nestable.js
@@ -20,7 +20,8 @@
     var hasTouch     = 'ontouchstart' in window,
         html         = $("html"),
         touchedlists = [],
-        $win         = UI.$win;
+        $win         = UI.$win,
+        draggingElement;
 
     /**
      * Detect CSS pointer-events property
@@ -143,14 +144,14 @@
                 }
                 e.preventDefault();
                 $this.dragStart(hasTouch ? e.touches[0] : e);
-                $this.trigger('nestable-start', [$this]);
+                $this.trigger('uk.nestable.start', [$this]);
             };
 
             var onMoveEvent = function(e) {
                 if ($this.dragEl) {
                     e.preventDefault();
                     $this.dragMove(hasTouch ? e.touches[0] : e);
-                    $this.trigger('nestable-move', [$this]);
+                    $this.trigger('uk.nestable.move', [$this]);
                 }
             };
 
@@ -158,8 +159,10 @@
                 if ($this.dragEl) {
                     e.preventDefault();
                     $this.dragStop(hasTouch ? e.touches[0] : e);
-                    $this.trigger('nestable-stop', [$this]);
+                    $this.trigger('uk.nestable.stop', [$this]);
                 }
+
+                draggingElement = false;
             };
 
             if (hasTouch) {
@@ -327,6 +330,8 @@
             this.dragEl = $(document.createElement(this.options.listNodeName)).addClass(this.options.listClass + ' ' + this.options.dragClass);
             this.dragEl.css('width', dragItem.width());
 
+            draggingElement = this.dragEl;
+
             this.tmpDragOnSiblings = [dragItem[0].previousSibling, dragItem[0].nextSibling];
 
             // fix for zepto.js
@@ -366,10 +371,10 @@
 
             if (this.tmpDragOnSiblings[0]!=el[0].previousSibling || this.tmpDragOnSiblings[0]!=el[0].previousSibling) {
 
-                this.element.trigger('nestable-change',[el, this.hasNewRoot ? "added":"moved"]);
+                this.element.trigger('uk.nestable.change',[el, this.hasNewRoot ? "added":"moved"]);
 
                 if (this.hasNewRoot) {
-                    this.dragRootEl.trigger('nestable-change', [el, "removed"]);
+                    this.dragRootEl.trigger('uk.nestable.change', [el, "removed"]);
                 }
             }
 
@@ -554,6 +559,22 @@
             }
         }
 
+    });
+
+    // adjust document scrolling
+    $('html').on('mousemove touchmove', function(e) {
+
+        if (draggingElement) {
+
+
+            var top = draggingElement.offset().top;
+
+            if (top < UI.$win.scrollTop()) {
+                UI.$win.scrollTop(UI.$win.scrollTop() - Math.ceil(draggingElement.height()/2));
+            } else if ( (top + draggingElement.height()) > (window.innerHeight + UI.$win.scrollTop()) ) {
+                UI.$win.scrollTop(UI.$win.scrollTop() + Math.ceil(draggingElement.height()/2));
+            }
+        }
     });
 
     // init code