+++ src/js/addons/search.js
@@ -65,7 +65,7 @@
                 $this.element.removeClass("uk-active");
             });
 
-            this.on('autocomplete-select', function(e, data) {
+            this.on('uk.autocomplete.select', function(e, data) {
                 if (data.url) {
                   location.href = data.url;
                 } else if(data.moreresults) {