+++ src/js/modal.js
@@ -22,6 +22,7 @@
             var $this = this;
 
             this.transition = UI.support.transition;
+            this.paddingdir = "padding-" + (UI.langdirection == 'left' ? "right":"left");
             this.dialog     = this.find(".uk-modal-dialog");
 
             this.on("click", ".uk-modal-close", function(e) {
@@ -84,19 +85,13 @@
 
         resize: function() {
 
-            var paddingdir = "padding-" + (UI.langdirection == 'left' ? "left":"right"),
-                margindir  = "margin-" + (UI.langdirection == 'left' ? "left":"right"),
-                bodywidth  = body.width();
+            var bodywidth  = body.width();
 
             this.scrollbarwidth = window.innerWidth - bodywidth;
 
-            $html.css(margindir, this.scrollbarwidth * -1);
+            body.css(this.paddingdir, this.scrollbarwidth);
 
-            this.element.css(paddingdir, "");
-
-            if (this.dialog.offset().left > this.scrollbarwidth) {
-                this.element.css(paddingdir, this.scrollbarwidth - (this.element[0].scrollHeight==window.innerHeight ? 0:this.scrollbarwidth ));
-            }
+            this.element.css('overflow-y', this.scrollbarwidth ? 'scroll' : 'auto');
 
             this.updateScrollable();
 
@@ -125,7 +120,9 @@
 
             this.element.hide().removeClass("uk-open");
 
-            $html.removeClass("uk-modal-page").css("margin-" + (UI.langdirection == 'left' ? "left":"right"), "");
+            $html.removeClass("uk-modal-page");
+
+            body.css(this.paddingdir, "");
 
             if(active===this) active = false;