+++ src/tests/_test.js
@@ -47,7 +47,6 @@
             "modal",
             "nav",
             "navbar",
-            "normalize",
             "offcanvas",
             "overlay",
             "pagination",
@@ -77,14 +76,14 @@
     $(function(){
 
 
-        var incustomizer = (iniframe && !window.parent.themes);
+        var incustomizer = (iniframe && !window.parent.themes), controls = $('<div class="uk-form uk-margin-top uk-margin-bottom"></div>');
 
         themes = $.extend(themes, window.parent.themes ? window.parent.themes:{});
         theme  = themes[theme] ? theme : 'Default';
 
         var testfolder = $("script[src$='_test.js']").attr("src").replace("_test.js", ""),
 
-            testselect = $('<select><option value="">- Select Test -</option><option value="overview.html">Overview</option></select>').css("margin", "20px 5px"),
+            testselect = $('<select><option value="">- Select Test -</option><option value="overview.html">Overview</option></select>').css("margin", "0 5px"),
             optgroup   = $('<optgroup label="Components"></optgroup>').appendTo(testselect);
 
 
@@ -98,7 +97,9 @@
                 if(testselect.val()) location.href = testfolder+testselect.val();
         });
 
-        $(".uk-container").prepend(testselect);
+        controls.prepend(testselect);
+
+        $(".uk-container:first").prepend(controls);
 
         if(incustomizer) {
             $("body").css("visibility", "visible");
@@ -146,9 +147,9 @@
                 var rtlcheckbox = $('<input type="checkbox">').on('change', function(e) {
                         localStorage['uikit.direction'] = ($(e.target).is(':checked') ? 'rtl' : 'ltr');
                         location.reload();
-                    }).css("margin", "20px 5px").prop('checked', direction == 'rtl'),
+                    }).css("margin", "0 5px").prop('checked', direction == 'rtl'),
 
-                    rtlcheckbox_label = $("<label>RTL mode</label>").css("margin", "20px 10px 20px 3px").prepend(rtlcheckbox);
+                    rtlcheckbox_label = $("<label>RTL mode</label>").prepend(rtlcheckbox);
 
                 if($.UIkit) $.UIkit.langdirection = rtlcheckbox.is(":checked") ? "right":"left";